Make diff report only files changed on this branch (#83)
I noticed this behavior when working with a very small pull request, but
which took a long time to review. For purposes of discussion, imagine
the following timeline:

* Monday, create a branch off of `main` and make a documentation change
(eg. update `README.md`)
* Tuesday, Wednesday, Thursday: other teams make commits unrelated to
mine, which get merged into the `main` branch
* In an effort to keep my branch up-to-date, I merge `main` into my
branch

After this sequence of events, cobra commander now reports that my
branch includes not only my change to `README.md`, but also *all of the
changes* from all of the other things merged into `main` since I
originally created my branch. This feels like it is overreporting the
scope of my PR, and will cause CI to test many additional components
that need not be tested.

This change proposes modifying the git check to only look for files
which are different on this PR vs. the base branch.